Move methods to be simple functions
